Create check_pull.yml
Added new workflow to check if more than 30 lines of code were changed. If so this will add a flag to signal that a pull request is needed. More work is needed in the setting of the repo to get this working. Tried multiple ways including adding branch rules and patterns.